infra/requirements.txt
boto3
aws-cdk.core
aws-cdk.aws-ec2
aws-cdk.aws-ecs
aws-cdk.aws-apigateway
aws-cdk.aws-iam
aws-cdk-lib
aws-cdk.aws-ecs-patterns